requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pymysql==1.1.0
websockets==12.0
google-generativeai==0.3.2
pydantic==2.5.0
pydantic-settings==2.1.0